Moved builder parameters to params module
