+++ src/api/checks.rs
@@ -1,5 +1,6 @@
 use crate::models::{CheckRunId, CheckSuiteId};
 use crate::params::repos::Commitish;
+use crate::params::checks::{CheckRunConclusion, CheckRunOutput, CheckRunStatus};
 use crate::{models, Octocrab, Result};
 use chrono::{DateTime, Utc};
 
@@ -10,72 +11,6 @@ pub struct ChecksHandler<'octo> {
     crab: &'octo Octocrab,
     owner: String,
     repo: String,
-}
-
-#[derive(serde::Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum CheckRunStatus {
-    Queued,
-    InProgress,
-    Completed,
-}
-
-#[derive(serde::Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum CheckRunConclusion {
-    Success,
-    Failure,
-    Neutral,
-    Cancelled,
-    TimedOut,
-    Skipped,
-    Stale,
-    ActionRequired
-}
-
-#[derive(serde::Serialize)]
-pub struct CheckRunOutput {
-    title: String,
-    summary: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    text: Option<String>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    annotations: Vec<CheckRunOutputAnnotation>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    images: Vec<CheckRunOutputImage>
-}
-
-#[derive(serde::Serialize)]
-pub struct CheckRunOutputAnnotation {
-    path: String,
-    start_line: u32,
-    end_line: u32,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    start_column: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    end_column: Option<u32>,
-    annotation_level: CheckRunOutputAnnotationLevel,
-    message: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    title: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    raw_details: Option<String>
-}
-
-#[derive(serde::Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum CheckRunOutputAnnotationLevel {
-    Notice,
-    Warning,
-    Failure
-}
-
-#[derive(serde::Serialize)]
-pub struct CheckRunOutputImage {
-    image_url: String,
-    alt: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    caption: Option<String>
 }
 
 #[derive(serde::Serialize)]
